requirements.txt
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.12.3
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.8
django-admin-numeric-filter==0.1.5
django-admin-rangefilter==0.5.1
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-mptt==0.10.0
django-taggit==1.1.0
html5lib==1.0.1
lazy-object-proxy==1.4.1
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
Pyphen==0.9.5
pytz==2019.2
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.0
WeasyPrint==50
webencodings==0.5.1
wrapt==1.11.2
